rename `/community/*` to `/api/*` to standardize it
